Move class only used for testing
